move rc_djb2 and rc_error_str to rc_util
