+++ src/rc_util.h
@@ -1,0 +1,14 @@
+#ifndef RC_UTIL_H
+#define RC_UTIL_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+unsigned rc_djb2(const char* input);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* RC_UTIL_H */
