+++ test/rcheevos-test.vcxproj
@@ -148,6 +148,7 @@
     <ClCompile Include="..\src\rc_client.c" />
     <ClCompile Include="..\src\rc_compat.c" />
     <ClCompile Include="..\src\rc_libretro.c" />
+    <ClCompile Include="..\src\rc_util.c" />
     <ClCompile Include="..\src\rhash\cdreader.c" />
     <ClCompile Include="..\src\rhash\hash.c" />
     <ClCompile Include="..\src\rhash\md5.c" />
@@ -231,6 +232,7 @@
     <ClInclude Include="..\src\rc_client_internal.h" />
     <ClInclude Include="..\src\rc_compat.h" />
     <ClInclude Include="..\src\rc_libretro.h" />
+    <ClInclude Include="..\src\rc_util.h" />
     <ClInclude Include="..\src\rc_version.h" />
     <ClInclude Include="..\src\rhash\md5.h" />
     <ClInclude Include="libretro.h" />
