+++ validator/validator.vcxproj
@@ -138,6 +138,7 @@
     <ClCompile Include="..\src\rcheevos\trigger.c" />
     <ClCompile Include="..\src\rcheevos\value.c" />
     <ClCompile Include="..\src\rc_compat.c" />
+    <ClCompile Include="..\src\rc_util.c" />
     <ClCompile Include="..\src\rhash\md5.c" />
     <ClCompile Include="..\test\lua\src\lapi.c" />
     <ClCompile Include="..\test\lua\src\lauxlib.c" />
