+++ validator/validator.vcxproj.filters
@@ -169,6 +169,9 @@
     <ClCompile Include="..\src\rc_compat.c">
       <Filter>src</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\rc_util.c">
+      <Filter>src</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\rcheevos.h">
